requirements/dev.txt
-r common.txt
Werkzeug==0.11.11
django-extensions==1.7.4
selenium==2.53.6
sqlparse==0.2.1
django-debug-toolbar==1.5
